:recycle: Small refactor, cleaning up helper funcs
